requirements.txt
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.8
idna==3.10
kaggle==1.6.17
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytrends==4.9.2
pytz==2025.1
requests==2.32.3
seaborn==0.13.2
six==1.17.0
text-unidecode==1.3
tqdm==4.67.1
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
Wikipedia-API==0.8.1